# src/openapi.yaml
openapi: 3.1.0
info:
  title: My brand - Eloi Chrysanthe
  description: An imaginary, but delightful Museum API for interacting with museum services and information. Built with love by Redocly.
  version: 1.0.0
  contact:
    email: eloi.chrysanthe@gmail.com
    url: "https://eloichrysanthe.me/"
servers:
  - url: http://127.0.0.1:20242/api
  - url: https://api.eloichrysanthe.me/api
components:
  securitySchemes:
    JWT:
      description: JWT bearer token description...
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /comments/one/{commentId}:
    patch:
      tags:
        - Comments
      summary: Update a comment
      operationId: updateComment
      parameters:
        - name: token
          in: header
          required: true
          schema:
            type: string
            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
        - name: commentId
          in: path
          required: true
          schema:
            type: string
          example: "65cbd47fcd78a22aaa14e812"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  description: This is to comment an article
                  example: this is a comment
      responses:
        "201":
          description: Successfully updated a comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: number
                    example: 201
        "400":
          description: "Validation Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 400
                  error:
                    type: array
                    items:
                      type: string
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: "Unauthorized: Missing invalid token."
...
  /comments/article/{articleId}/{commentId}:
    delete:
      tags:
        - Comments
      summary: Delete comment on an article
      operationId: deleteComment
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
          example: "65cbd47fcd78a22aaa14e812"
        - name: commentId
          in: path
          required: true
          schema:
            type: string
          example: "65cbd47fcd78a22aaa14e812"
        - name: token
          in: header
          required: true
          schema:
            type: string
            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
      responses:
        "200":
          description: Successfully fetched comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: number
                    example: 201
        "400":
          description: "Validation Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 400
                  error:
                    type: array
                    items:
                      type: string
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: "Unauthorized: Missing invalid token."
...
security:
  - JWT: []